Refactor contact info models
