+++ epp-client/src/contact.rs
@@ -1,3 +1,4 @@
 pub mod check;
 pub mod create;
 pub mod delete;
+pub mod info;
