+++ epp-client/src/epp.rs
@@ -5,7 +5,6 @@ pub mod request;
 pub mod response;
 pub mod xml;
 
-pub use request::contact::info::*;
 pub use request::contact::update::*;
 pub use request::host::check::*;
 pub use request::host::create::*;
@@ -15,7 +14,6 @@ pub use request::host::update::*;
 pub use request::message::ack::*;
 pub use request::message::poll::*;
 
-pub use response::contact::info::*;
 pub use response::contact::update::*;
 pub use response::host::check::*;
 pub use response::host::create::*;
