+++ epp-client/src/epp/request/contact.rs
@@ -1,4 +1,3 @@
 //! Types for EPP contact requests
 
-pub mod info;
 pub mod update;
