+++ epp-client/src/epp/response/contact.rs
@@ -1,4 +1,3 @@
 //! Types for EPP contact responses
 
-pub mod info;
 pub mod update;
