+++ epp-client/src/tests/se.rs
@@ -6,6 +6,7 @@ mod request {
     use crate::contact::check::EppContactCheck;
     use crate::contact::create::EppContactCreate;
     use crate::contact::delete::EppContactDelete;
+    use crate::contact::info::EppContactInfo;
     use crate::domain::check::EppDomainCheck;
     use crate::domain::create::EppDomainCreate;
     use crate::domain::delete::EppDomainDelete;
